Rewrites the readme
After thinking over the idea more carefully, I decided the readme needed to be rewritten. I took care this time to emphasize the experimental nature of this project and to focus on the more theoretical problem that it is attempting to solve at this point, as opposed to the more practical problems that were described previously (which are really incidental to the project at this point). I also detailed the tentative method of calculating relative credibility scores.